+++ test/cider-find-tests.el
@@ -1,6 +1,6 @@
 ;;; cider-find-tests.el  -*- lexical-binding: t; -*-
 
-;; Copyright © 2012-2024 Bozhidar Batsov
+;; Copyright © 2012-2025 Bozhidar Batsov
 
 ;; Author: Bozhidar Batsov <bozhidar@batsov.dev>